The final PharmacoSet object containing the multiple parts of the NCI-ALMANAC dataset.

The final PharmacoSet object containing the multiple parts of the NCI-ALMANAC dataset.

Usage

data(NCI_ALMANAC_pset)

CBWWorkshop2024::NCI_ALMANAC_pset

Format

PharmacoSet object

sample

60 samples

treatment

31 treatments

molecularProfiles

A MultiAssayExperiment with 4 RangedSummarizedExperiments, 'rna', 'mirna', 'rnaseq.comp', and 'rnaseq.iso'

treatmentResponse

A TreatmentResponseExperiment with 10080 total experiments. Assays are: 'response', 'profiles' and 'assay_metadata'.