A data.frame containing cell-line metadata manually curated as part of the PharmacoSet creation process for the NCI-ALMANAC.

Sample level metadata on cell-lines from the NCI-ALMANAC, standardized to use identifers from Cellosaurus and included manually curated tissue annotations from Oncotree as well as Cellosaurus disease type annotations.

Treatment metadata on the various drugs used in the NCI-ALMANAC with additional annotations included standardized drug identifiers which match the nomenclature of other PharmacoSet object as well as links structural idenfiers such as SMILES and InchiKeys as well as links to PubChem via the cid.

Usage

data(NCI_ALMANAC_sample_metadata)

data(NCI_ALMANAC_sample_metadata)

Format

data.frame object

data.frame object

References

Bairoch A. he Cellosaurus, a cell line knowledge resource. J. Biomol. Tech. 29:25-38(2018) DOI: 10.7171/jbt.18-2902-002; PMID: 29805321

Holbeck, S. L., Camalier, R., Crowell, J. A., Govindharajulu, J. P., Hollingshead, M., Anderson, L. W., Polley, E., Rubinstein, L., Srivastava, A., Wilsker, D., Collins, J. M., & Doroshow, J. H. (2017). The National Cancer Institute ALMANAC: A Comprehensive Screening Resource for the Detection of Anticancer Drug Pairs with Enhanced Therapeutic Activity. Cancer Research, 77(13), 3564–3576. https://doi.org/10.1158/0008-5472.CAN-17-0489

Holbeck, S. L., Camalier, R., Crowell, J. A., Govindharajulu, J. P., Hollingshead, M., Anderson, L. W., Polley, E., Rubinstein, L., Srivastava, A., Wilsker, D., Collins, J. M., & Doroshow, J. H. (2017). The National Cancer Institute ALMANAC: A Comprehensive Screening Resource for the Detection of Anticancer Drug Pairs with Enhanced Therapeutic Activity. Cancer Research, 77(13), 3564–3576. https://doi.org/10.1158/0008-5472.CAN-17-0489